Improve percentage, perform inversion on equal key
For percentage mode of multiplication and division keep the first
oeprand as the constant (the base), and treat the equal command as the
request to perform inversion. For example, 200 * 5 % gives 20, then
entering 6 % gives 12, here 200 is the base. Then, 200 * 5 % gives 20
and further pressing = gives 5 back. Then, pressing 6 % gives 12 and
further pressing = gives 6 back. The same is for division, 10 / 5 %
gives 200 (which means, 10 is 5 % of 200), then pressing = gives 5 back,
and so on.